+++ lib/src/visual/views/chats/chat_messages_components/entries/chat_flame_restored.entry.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:twonly/src/database/twonly.db.dart';
 import 'package:twonly/src/model/protobuf/client/generated/data.pb.dart';
 import 'package:twonly/src/utils/misc.dart';
+import 'package:twonly/src/visual/components/animate_icon.comp.dart';
 import 'package:twonly/src/visual/elements/better_text.element.dart';
 
 import 'package:twonly/src/visual/views/chats/chat_messages_components/entries/common.dart';
@@ -42,14 +43,27 @@ class ChatFlameRestoredEntry extends StatelessWidget {
       ),
       padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 6),
       decoration: BoxDecoration(
-        color: Colors.orange,
+        color: info.color,
         borderRadius: borderRadius,
       ),
-      child: BetterText(
-        text: context.lang.chatEntryFlameRestored(
-          data.restoredFlameCounter.toInt(),
-        ),
-        textColor: isDarkMode(context) ? Colors.black : Colors.black,
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const SizedBox(
+            width: 24,
+            height: 24,
+            child: EmojiAnimationComp(emoji: '🔥'),
+          ),
+          const SizedBox(width: 8),
+          Flexible(
+            child: BetterText(
+              text: context.lang.chatEntryFlameRestored(
+                data.restoredFlameCounter.toInt(),
+              ),
+              textColor: isDarkMode(context) ? Colors.black : Colors.black,
+            ),
+          ),
+        ],
       ),
     );
   }