+++ test/mocks/platform_channels.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: avoid_print, avoid_dynamic_calls
+// ignore_for_file: avoid_dynamic_calls
 
 import 'dart:async';
 import 'package:flutter/services.dart';