+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New: Adds an "Ask a Friend" button to new contact suggestions.
 - New: Adds security profiles.
 - Improved: Onboarding flow for new users.
+- Improved: Flame restore experience.
 - Improved: The blue verification checkmark now displays the total number of verifications.
 - Fix: Issue with receiving messages when user closed app while decrypting
 - Fix: Background message fetching reliability.